Refactor out some router calculations with view of supporting arrival-based routing
