+++ shared/src/commonMain/kotlin/cl/emilym/sinatra/router/Raptor.kt
@@ -22,14 +22,10 @@ data class RaptorConfig(
 )
 
 class Raptor(
-    private val graph: NetworkGraph,
-    activeServices: List<List<ServiceId>>,
-    private val config: RaptorConfig
-) {
-
-    private val activeServices = activeServices.map {
-        it.associate { graph.mappings.serviceIds.indexOf(it) to Unit }
-    }
+    override val graph: NetworkGraph,
+    override val activeServiceIds: List<List<ServiceId>>,
+    override val config: RaptorConfig
+): BaseRouter() {
 
     fun calculate(
         departureTime: DaySeconds,
@@ -48,201 +44,34 @@ class Raptor(
         departureStops: List<RaptorStop>,
         arrivalStops: List<RaptorStop>
     ): RaptorJourney {
-        val arrivalStopIndices = arrivalStops.mapNotNull { stop -> graph.mappings.stopIdToIndex[stop.id]?.let { it to stop } }
-        val departureStopIndices = departureStops.mapNotNull { stop -> graph.mappings.stopIdToIndex[stop.id]?.let { it to stop } }
-        if (arrivalStopIndices.isEmpty() || departureStopIndices.isEmpty()) {
-            throw RouterException.stopNotFound()
-        }
+        return doCalculation(
+            departureTime,
+            departureStops,
+            arrivalStops
+        )
+    }
 
-        val nodeCount = graph.metadata.nodeCount.toInt()
-
-        // Dijkstra
-        val Q = FibonacciHeap<Int,Long>()
-        val checked = mutableMapOf<Int, Unit>()
-        val dist = Array(nodeCount) { Long.MAX_VALUE }
-        val distP = Array(nodeCount) { Long.MAX_VALUE }
-        val prev = Array<Int?>(nodeCount) { null }
-        val prevEdge = Array<NetworkGraphEdge?>(nodeCount) { null }
-        val dayIndex = Array<Int?>(nodeCount) { null }
-
+    override fun initializeDjikstra(
+        Q: FibonacciHeap<Int, Long>,
+        dist: Array<Long>,
+        distP: Array<Long>,
+        departureStopIndices: List<Pair<Int, RaptorStop>>,
+        arrivalStopIndices: List<Pair<Int, RaptorStop>>
+    ) {
         for (departureStopIndex in departureStopIndices) {
             dist[departureStopIndex.first] = departureStopIndex.second.addedTime
             distP[departureStopIndex.first] = departureStopIndex.second.addedTime
             Q.add(departureStopIndex.first, 0)
         }
-
-        while (!Q.isEmpty()) {
-            val u = Q.pop()!!
-            if (checked[u] != null) continue
-            checked[u] = Unit
-
-            val neighbours = getNeighbours(
-                u,
-                departureTime + dist[u],
-                prevEdge[u]?.type == EdgeType.TRANSFER
-            )
-
-            for (neighbour in neighbours) {
-                val v = neighbour.index
-                val altP = distP[u] + neighbour.costP
-                val alt = dist[u] + neighbour.cost
-                if (altP < distP[v]) {
-                    prev[v] = u
-                    prevEdge[v] = neighbour.edge
-                    distP[v] = altP
-                    dist[v] = alt
-                    dayIndex[v] = neighbour.dayIndex
-                    Q.add(v, altP)
-                    checked.remove(v)
-                }
-                if (neighbour.edge.type == EdgeType.TRAVEL) {
-                    val tV = getNode(neighbour.edge.connectedNodeIndex.toInt()).stopIndex.toInt()
-                    val addedPenalty = config.changeOverPenalty
-                    val addedTime = config.changeOverTime
-                    if ((altP + addedPenalty) < distP[tV]) {
-                        prev[tV] = u
-                        prevEdge[tV] = neighbour.edge
-                        distP[tV] = altP + addedPenalty
-                        dist[tV] = alt + addedTime
-                        dayIndex[v] = neighbour.dayIndex
-                        Q.add(tV, altP)
-                        checked.remove(v)
-                    }
-                }
-            }
-        }
-
-        if (arrivalStopIndices.all { prev[it.first] == null }) throw RouterException.noJourneyFound()
-
-        val options = arrivalStopIndices
-            .filterNot { prev[it.first] == null }
-
-        // Reconstruct journey
-        var cursor = (
-            options.filter {
-                prevEdge[it.first]?.type != EdgeType.TRANSFER ||
-                        (prevEdge[it.first]?.type == EdgeType.TRANSFER &&
-                                (dist[it.first] + it.second.addedTime) >= config.maximumWalkingTime)
-            }.nullIfEmpty()
-        )?.minBy { dist[it.first] + it.second.addedTime }?.first ?: throw RouterException.noJourneyFound()
-
-        val chain = mutableListOf<RaptorJourneyConnection>()
-        var connection: RaptorJourneyConnection? = null
-        var edgeType: EdgeType? = null
-
-        fun addConnection(c: RaptorJourneyConnection) {
-            val c = when (c) {
-                is RaptorJourneyConnection.Travel -> c.copy(
-                    stops = listOf(graph.mappings.stopIds[getNode(cursor).stopIndex.toInt()]) + c.stops
-                )
-                is RaptorJourneyConnection.Transfer -> c.copy(
-                    stops = listOf(graph.mappings.stopIds[getNode(cursor).stopIndex.toInt()]) + c.stops
-                )
-            }
-            chain.add(0, c)
-        }
-
-        while (cursor !in departureStopIndices.map { it.first }) {
-            val edge = prevEdge[cursor]!!
-            val node = getNode(edge.connectedNodeIndex.toInt())
-
-            if (edgeType != edge.type) {
-                if (connection != null) addConnection(connection)
-                edgeType = edge.type
-                connection = when (edge.type) {
-                    EdgeType.TRAVEL -> {
-                        RaptorJourneyConnection.Travel(
-                            listOf(),
-                            graph.mappings.routeIds[node.routeIndex.toInt()],
-                            graph.mappings.headings[node.headingIndex.toInt()],
-                            0,
-                            edge.departureTime.toLong() + edge.cost.toLong(),
-                            dayIndex[cursor] ?: 0,
-                            0
-                        )
-                    }
-                    EdgeType.TRANSFER, EdgeType.TRANSFER_NON_ADJUSTABLE -> {
-                        RaptorJourneyConnection.Transfer(
-                            listOf(),
-                            0
-                        )
-                    }
-                    else -> null
-                }
-            }
-
-            connection = when(edge.type) {
-                EdgeType.TRAVEL -> {
-                    val c = (connection as RaptorJourneyConnection.Travel)
-                    c.copy(
-                        stops = listOf(graph.mappings.stopIds[node.stopIndex.toInt()]) + c.stops,
-                        startTime = edge.departureTime.toLong(),
-                        travelTime = c.endTime - edge.departureTime.toLong(),
-                        dayIndex = dayIndex[cursor] ?: 0
-                    )
-                }
-                EdgeType.TRANSFER, EdgeType.TRANSFER_NON_ADJUSTABLE -> {
-                    val c = (connection as RaptorJourneyConnection.Transfer)
-                    c.copy(
-                        stops = listOf(graph.mappings.stopIds[node.stopIndex.toInt()]) + c.stops,
-                        travelTime = c.travelTime + edge.cost.toLong()
-                    )
-                }
-                else -> null
-            }
-
-            cursor = prev[cursor]!!
-        }
-
-        if (connection != null) addConnection(connection)
-
-        return RaptorJourney(chain)
     }
 
-    private fun getNode(index: NodeIndex) = graph.node(index)
+    override fun isValidTravelEdge(
+        edge: NetworkGraphEdge,
+        dayAdjustment: Seconds,
+        anchorTime: DaySeconds
+    ): Boolean = (edge.departureTime.toLong() + dayAdjustment + 60) < anchorTime
 
-    private fun getNeighbours(index: NodeIndex, departureTime: DaySeconds, ignoreTransfer: Boolean = false): List<NodeCost> {
-        val node = graph.node(index)
-        val edges = node.edges
-
-        return edges.flatMap {
-            when (it.type) {
-                EdgeType.UNWEIGHTED -> listOf(
-                    NodeCost(it.connectedNodeIndex.toInt(), 0L, 0L, it, null)
-                )
-                EdgeType.TRANSFER, EdgeType.TRANSFER_NON_ADJUSTABLE -> {
-                    if (ignoreTransfer) return@flatMap emptyList()
-                    if (it.cost.toLong() > (config.maximumWalkingTime)) return@flatMap emptyList()
-                    listOf(NodeCost(
-                        it.connectedNodeIndex.toInt(),
-                        it.cost.toLong() + config.transferTime,
-                        it.cost.toLong() + config.transferPenalty,
-                        it,
-                        null
-                    ))
-                }
-                EdgeType.TRAVEL -> {
-                    val daysActive = (-1..1).filter { d -> servicesAreActive(it.availableServices, d) }
-                    if (daysActive.isEmpty()) return@flatMap emptyList()
-                    daysActive.mapNotNull { d ->
-                        val segmentDepartureTime = it.departureTime.toInt() + (86400 * d)
-                        if ((segmentDepartureTime + 60) < departureTime) return@mapNotNull null
-                        val cost = (segmentDepartureTime - departureTime) + it.cost.toLong()
-                        NodeCost(it.connectedNodeIndex.toInt(), cost, cost, it, d)
-                    }
-                }
-            }
-        }
-    }
-
-    private fun servicesAreActive(
-        serviceIndices: List<ServiceIndex>,
-        day: Int
-    ): Boolean {
-        for (i in serviceIndices) {
-            if (activeServices[day + 1].containsKey(i.toInt())) return true
-        }
-        return false
-    }
+    override fun getTravelAnchorTime(edge: NetworkGraphEdge, dayAdjustment: Seconds): Long =
+        (edge.departureTime.toLong() + dayAdjustment)
 
 }
